+++ packages/af-utils-react-virtual-headless/src/hooks/useVirtual/index.js
@@ -1,6 +1,6 @@
 import { useLayoutEffect, useEffect } from "react";
 import useVirtualModel from "../useVirtualModel";
-import { DEFAULT_OVERSCAN_COUNT } from "constants";
+import { DEFAULT_OVERSCAN_COUNT, DEFAULT_ESTIMATED_ITEM_SIZE } from "constants";
 
 const localUseEffect = process.env.__IS_SERVER__ ? useEffect : useLayoutEffect;
 
@@ -11,7 +11,10 @@ const useVirtual = params => {
         model._startBatch();
         model.setOverscan(params.overscanCount ?? DEFAULT_OVERSCAN_COUNT);
         model.setHorizontal(!!params.horizontal);
-        model.setItemCount(params.itemCount, params.getEstimatedItemSize);
+        model.setItemCount(
+            params.itemCount,
+            params.estimatedItemSize || DEFAULT_ESTIMATED_ITEM_SIZE
+        );
         model._endBatch();
     });