+++ packages/af-utils-react-virtual-headless/src/hooks/useVirtualModel/index.js
@@ -1,14 +1,15 @@
 import List from "models/List";
 import {
     DEFAULT_OVERSCAN_COUNT,
-    DEFAULT_ESTIMATED_WIDGET_SIZE
+    DEFAULT_ESTIMATED_WIDGET_SIZE,
+    DEFAULT_ESTIMATED_ITEM_SIZE
 } from "constants";
 
 import useOnce from "../useOnce";
 
 const useVirtualModel = ({
     itemCount = 0,
-    getEstimatedItemSize,
+    estimatedItemSize = DEFAULT_ESTIMATED_ITEM_SIZE,
     estimatedWidgetSize = DEFAULT_ESTIMATED_WIDGET_SIZE,
     overscanCount = DEFAULT_OVERSCAN_COUNT,
     horizontal
@@ -18,7 +19,7 @@ const useVirtualModel = ({
 
         model.setOverscan(overscanCount);
         model.setHorizontal(!!horizontal);
-        model.setItemCount(itemCount, getEstimatedItemSize);
+        model.setItemCount(itemCount, estimatedItemSize);
         model.setWidgetSize(estimatedWidgetSize);
 
         return model;