+++ packages/af-utils-react-table/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @af-utils/react-table
 
+## 0.0.39
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.0.34
+
 ## 0.0.38
 
 ### Patch Changes